Refactor the code
