+++ src/project.js
@@ -1,18 +1,11 @@
-const { filteringBtns } = require("./filter");
+import { filteringBtns } from './filter';
+import { updateProjects, updatedProjects, validateInput } from './utils';
 
 const projectBtn = document.querySelector('#project-btn');
 const projectName = document.querySelector('#new-project');
 
 const createProject = (name) => {
   return { name };
-}
-
-const updatedProjects = () => {
-  return JSON.parse(localStorage.getItem('projects'));
-}
-
-const updateProjects = (arr) => {
-  localStorage.setItem('projects', JSON.stringify(arr));
 }
 
 const displayProjects = () => {
@@ -29,13 +22,6 @@ const addNewProject = (name) => {
   temp.push(item);
   updateProjects(temp);
   projectName.value = "";
-}
-
-const validateInput = (id) => {
-  if(id.value === null || id.value === "") {
-    alert('Please fill the required fields');
-    return true;
-  }
 }
 
 const addProjectsToSelection = () => {
@@ -58,14 +44,7 @@ const projectCreation = () => {
   });
 }
 
-const projectTags = document.querySelector('p')
-
-for(let pTags in projectTags) {
-  const delBtn = docum
-}
-
-
-module.exports = {
+export {
   displayProjects,
   projectCreation,
   updatedProjects,
